requirements.txt
Flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.1
python-dotenv==1.0.0
langdetect==1.0.9
pytz==2024.1
